Prototype the getPlaces query for the app #apollo-client
[UI]

Start the api server. Use [graphiql](http://localhost:8080/graphiql) to build a query that will return a result like this (be sure to include all the fields):

```
{
  "data": {
    "places": [
      {
        "id": "5b4f6653129119bff0c100df",
        "name": "San Francisco",
        "visited": true,
        "location": {
          "latitude": 37.7749295,
          "longitude": -122.4194155,
          "weather": {
            "icon": "☀️",
            "temperature": 67.35
          }
        }
      },
      ...
    ]
  }
}
```

When you have a working query, save it in `ui/src/graphql/getPlacesQuery.js` (there is a file there waiting). Be sure to enable/uncomment the `gql` import